+++ src/main/java/org/flowsoft/flowg/nodes/functions/GCodeFuncNode.java
@@ -1,8 +1,8 @@
-package org.flowsoft.flowg.nodes;
+package org.flowsoft.flowg.nodes.functions;
 
+import org.flowsoft.flowg.nodes.IdentifierNode;
 import org.flowsoft.flowg.nodes.base.Node;
 import org.flowsoft.flowg.nodes.base.StatementNode;
-import org.flowsoft.flowg.nodes.functions.FormalParameterListNode;
 import org.flowsoft.flowg.visitors.IVisitor;
 
 public class GCodeFuncNode extends Node implements StatementNode {
