+++ src/main/java/org/flowsoft/flowg/symboltables/SymbolTable.java
@@ -2,7 +2,7 @@ package org.flowsoft.flowg.symboltables;
 
 import org.flowsoft.flowg.Type;
 import org.flowsoft.flowg.TypeException;
-import org.flowsoft.flowg.nodes.GCodeCodeNode;
+import org.flowsoft.flowg.nodes.functions.GCodeCodeNode;
 import org.flowsoft.flowg.nodes.functions.FormalParameterNode;
 import org.flowsoft.flowg.nodes.StatementListNode;
 import org.flowsoft.flowg.Cloneable;
