+++ src/main/java/org/flowsoft/flowg/visitors/CodeGeneratingVisitor.java
@@ -16,7 +16,6 @@ import java.math.BigDecimal;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.BiFunction;
-import java.util.regex.MatchResult;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
